refactor(comparateur): déplace les anciens composants dans choix-du-statut
