+++ site/source/locales/ui-fr.yaml
@@ -841,6 +841,35 @@ pages:
           choisir le statut juridique le plus adapté à votre projet
           d'entreprise.
         title: Aide au choix du statut juridique
+      modifier-options:
+        acre:
+          body: L’aide à la création ou à la reprise d’une entreprise (Acre) consiste en
+            une <2>exonération partielle de charges sociales</2>, dite
+            exonération de début d’activité, <4>pendant 12 mois</4>.
+          info-title: Bénéficier de l’Acre
+          label: Activer l’Acre dans la simulation
+          title: Acre
+        confirm-label: Enregistrer les options
+        imposition:
+          body: Vous pouvez <2>choisir entre l’imposition sur les sociétés et sur le
+            revenu</2> durant les 5 premières années.
+          body-ae: En tant qu’auto-entreprise (AE), c’est l’<1>impôt sur le revenu</1> qui
+            est appliqué automatiquement. Dans certaines situations, vous pouvez
+            aussi opter pour le <4><0>versement libératoire</0></4>.
+          hors-ae:
+            disclaimer: À ce jour, ce comparateur ne prend pas en compte le calcul de
+              l’impôt sur le revenu pour les SAS(U).
+            label: Mode d’imposition (hors AE)
+          title: Imposition
+          versement-libératoire:
+            title: Versement libératoire (pour AE)
+        title: Modifier mes options
+        versement-libératoire:
+          label: Activer le versement libératoire dans la simulation
+      revenu-table:
+        net: Tableau affichant le détail du calcul du revenu net pour chaque statut
+        net-après-impôt: Tableau affichant le détail du calcul du revenu net après impôt
+          pour chaque statut
       rémunération:
         CA:
           help:
@@ -1791,31 +1820,6 @@ pages:
         description: Auto-entrepreneur, EI/EURL ou SASU ? Simulez les différences de
           revenu, retraite et indemnités maladie
         title: Comparateur de statut juridique
-      modifier-options:
-        acre:
-          body: L’aide à la création ou à la reprise d’une entreprise (Acre) consiste en
-            une <2>exonération partielle de charges sociales</2>, dite
-            exonération de début d’activité, <4>pendant 12 mois</4>.
-          info-title: Bénéficier de l’Acre
-          label: Activer l’Acre dans la simulation
-          title: Acre
-        confirm-label: Enregistrer les options
-        imposition:
-          body: Vous pouvez <2>choisir entre l’imposition sur les sociétés et sur le
-            revenu</2> durant les 5 premières années.
-          body-ae: En tant qu’auto-entreprise (AE), c’est l’<1>impôt sur le revenu</1> qui
-            est appliqué automatiquement. Dans certaines situations, vous pouvez
-            aussi opter pour le <4><0>versement libératoire</0></4>.
-          hors-ae:
-            disclaimer: À ce jour, ce comparateur ne prend pas en compte le calcul de
-              l’impôt sur le revenu pour les SAS(U).
-            label: Mode d’imposition (hors AE)
-          title: Imposition
-          versement-libératoire:
-            title: Versement libératoire (pour AE)
-        title: Modifier mes options
-        versement-libératoire:
-          label: Activer le versement libératoire dans la simulation
       montants:
         CA: Chiffre d’affaires estimé
         imposition:
@@ -1867,10 +1871,6 @@ pages:
         impôt:
           méthode: Méthode de calcul de l’impôt sur le revenu
           taux: Taux d’imposition
-      revenu-table:
-        net: Tableau affichant le détail du calcul du revenu net pour chaque statut
-        net-après-impôt: Tableau affichant le détail du calcul du revenu net après impôt
-          pour chaque statut
       réponses:
         activité:
           artisanale: Artisanale
