+++ site/source/pages/assistants/choix-du-statut/_components/RevenuTable.tsx
@@ -19,7 +19,7 @@ export default function RevenuTable({
 				<StyledTable>
 					<caption className="sr-only">
 						{t(
-							'pages.simulateurs.comparaison-statuts.revenu-table.net',
+							'pages.assistants.choix-statut.revenu-table.net',
 							'Tableau affichant le détail du calcul du revenu net pour chaque statut'
 						)}
 					</caption>
@@ -131,7 +131,7 @@ export default function RevenuTable({
 				<StyledTable>
 					<caption className="sr-only">
 						{t(
-							'pages.simulateurs.comparaison-statuts.revenu-table.net-après-impôt',
+							'pages.assistants.choix-statut.revenu-table.net-après-impôt',
 							'Tableau affichant le détail du calcul du revenu net après impôt pour chaque statut'
 						)}
 					</caption>
